Add ability to stop a currently running cell.
