+++ src/vs/workbench/contrib/positronNotebook/browser/PositronNotebookComponent.css
@@ -22,74 +22,3 @@
 	flex-direction: column;
 	gap: 1rem;
 }
-
-.positron-notebook-cell {
-	outline: 1px solid var(--vscode-focusBorder, pink);
-	/* padding-inline: 1rem;
-	padding-block: 0.25rem; */
-	border-radius: 0.25rem;
-	position: relative;
-
-	.run-button {
-		position: absolute;
-		top: 0.25rem;
-		right: 0.25rem;
-		aspect-ratio: 1;
-		display: grid;
-		place-content: center;
-		padding: 2px;
-
-		&:hover {
-			scale: 1.2;
-
-			.button-icon::before {
-				color: blue;
-			}
-		}
-
-		.button-icon {
-			align-items: center;
-			justify-content: center;
-			display: flex;
-			width: 16px;
-			aspect-ratio: 1;
-		}
-
-		.button-icon::before {
-			color: inherit;
-		}
-
-		&.disabled {
-			color: grey;
-		}
-	}
-
-	.positron-notebook-cell-code {
-		background-color: var(--vscode-notebook-cellEditorBackground, white);
-	}
-
-	.positron-notebook-cell-outputs {
-		background-color: var(
-			--vscode-notebook-outputContainerBackgroundColor,
-			white
-		);
-	}
-
-	.positron-notebook-cell-code,
-	.positron-notebook-cell-outputs {
-		padding-inline: 1rem;
-		padding-block: 0.5rem;
-	}
-
-	& pre {
-		margin: 0;
-	}
-
-	.execution-status {
-		position: absolute;
-		right: 3px;
-		bottom: 0;
-		font-size: 0.7rem;
-		font-style: italic;
-	}
-}
