+++ src/vs/workbench/contrib/positronNotebook/browser/PositronNotebookComponent.tsx
@@ -6,14 +6,12 @@ import 'vs/css!./PositronNotebookComponent';
 
 import * as React from 'react';
 import { ISize } from 'vs/base/browser/positronReactRenderer';
-import { PositronButton } from 'vs/base/browser/ui/positronComponents/positronButton';
 import { ISettableObservable } from 'vs/base/common/observableInternal/base';
 import { NotebookCellTextModel } from 'vs/workbench/contrib/notebook/common/model/notebookCellTextModel';
 import { InputObservable } from 'vs/workbench/contrib/positronNotebook/browser/PositronNotebookEditor';
 import { CellExecutionStatusCallback, NotebookKernelObservable, NotebookViewModelObservable } from 'vs/workbench/contrib/positronNotebook/browser/PositronNotebookWidget';
+import { NotebookCell } from './NotebookCell';
 import { useObservedValue } from './useObservedValue';
-import { gatherOutputContents } from 'vs/workbench/contrib/positronNotebook/browser/getOutputContents';
-import { NotebookCellExecutionState } from 'vs/workbench/contrib/notebook/common/notebookCommon';
 
 type CellsExecutionCallback = (cells?: Iterable<NotebookCellTextModel>) => Promise<void>;
 
@@ -44,7 +42,7 @@ export function PositronNotebookComponent(
 			</div>
 			<div className='positron-notebook-cells-container'>
 				<h2>Notebook Cells</h2>
-				{notebookCells.map(cell => <CellDisplay
+				{notebookCells.map(cell => <NotebookCell
 					key={cell.handle}
 					onRunCell={() => executeCells([cell])}
 					getCellExecutionStatus={getCellExecutionStatus}
@@ -56,92 +54,3 @@ export function PositronNotebookComponent(
 }
 
 
-type ExecutionStateString = 'running' | 'pending' | 'unconfirmed' | 'idle';
-function parseExecutionState(state?: NotebookCellExecutionState): ExecutionStateString {
-	switch (state) {
-		case NotebookCellExecutionState.Executing:
-			return 'running';
-		case NotebookCellExecutionState.Pending:
-			return 'pending';
-		case NotebookCellExecutionState.Unconfirmed:
-			return 'unconfirmed';
-		default:
-			return 'idle';
-	}
-}
-
-function useRunCell({ cell, onRunCell, getCellExecutionStatus }: {
-	cell: NotebookCellTextModel;
-	onRunCell: () => Promise<void>;
-	getCellExecutionStatus: CellExecutionStatusCallback;
-}) {
-
-	const [executionStatus, setExecutionStatus] = React.useState<ExecutionStateString>('idle');
-
-	const runCell = React.useCallback(() => {
-		setExecutionStatus('running');
-		onRunCell()
-			.then(() => {
-				setExecutionStatus('idle');
-			}).catch(() => {
-				setExecutionStatus(parseExecutionState(getCellExecutionStatus(cell)?.state));
-			});
-	}, [onRunCell, getCellExecutionStatus, cell]);
-
-
-	return {
-		runCell,
-		executionStatus,
-	};
-}
-
-function CellDisplay({ cell, onRunCell, getCellExecutionStatus }: {
-	cell: NotebookCellTextModel;
-	onRunCell: () => Promise<void>;
-	getCellExecutionStatus: CellExecutionStatusCallback;
-}) {
-
-	const outputContents = useCellOutputContents(cell);
-
-	const {
-		runCell,
-		executionStatus
-	} = useRunCell({ cell, onRunCell, getCellExecutionStatus });
-
-	return (
-		<div className='positron-notebook-cell'>
-			<PositronButton className='run-button' ariaLabel='Run cell' onClick={runCell}>
-				<div className={`button-icon codicon codicon-run`} />
-			</PositronButton>
-			<pre className='positron-notebook-cell-code'>{cell.getValue()}</pre>
-			<div className='positron-notebook-cell-outputs'>
-				{outputContents ? outputContents.map(({ content, id }) => <div key={id}>{content}</div>) : 'No outputs'}
-			</div>
-			<div className='execution-status'>Status: {executionStatus}</div>
-		</div>
-	);
-}
-
-
-/**
- * Gather contents of output for a cell and update them as they change.
- * @param cell A notebook cell
- * @returns An array of output contents that are updated as the cell trigers the
- * `onDidChangeOutputs` event.
- */
-function useCellOutputContents(cell: NotebookCellTextModel) {
-
-	const [outputContents, setOutputContents] = React.useState(
-		gatherOutputContents(cell)
-	);
-
-	React.useEffect(() => {
-		const outputListener = cell.onDidChangeOutputs(() => {
-			setOutputContents(gatherOutputContents(cell));
-		});
-
-		return () => outputListener.dispose();
-	}, [cell]);
-
-	return outputContents;
-}
